+++ crates/web-client/test/wallet.test.ts
@@ -2,7 +2,7 @@ import { createNewWallet, isValidAddress } from "./webClientTestUtils.js";
 
 describe("wallet tests", () => {
   it("create a new wallet", async () => {
-    const result = await createNewWallet("OffChain", false);
+    const result = await createNewWallet("Private", false);
 
     isValidAddress(result);
   });